Extract dependency chain to a separate package
